\name{vcf2eqtl-package}
\alias{vcf2eqtl-package}
\alias{vcf2eqtl}
\docType{package}
\title{
\packageTitle{vcf2eqtl}
}
\description{
\packageDescription{vcf2eqtl}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vcf2eqtl}
\packageIndices{vcf2eqtl}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{vcf2eqtl}

Maintainer: \packageMaintainer{vcf2eqtl}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
